metadatamanager/requirements.txt
Django>=4.2.1
psycopg2>=2.9.2
django-moderation
